make it possible to run a non-sigleton version of yargs
